CASMCMS-8553: Update BOS tests to handle tenant fields in the responses from BOS
